+++ androidutilities-room/src/main/java/com/heyzeusv/androidutilities/room/RoomProcessor.kt
@@ -8,7 +8,9 @@ import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import com.google.devtools.ksp.validate
-import com.heyzeusv.androidutilities.room.csv.CsvConverterCreator
+import com.heyzeusv.androidutilities.room.creators.CsvConverterCreator
+import com.heyzeusv.androidutilities.room.creators.EntityFilesCreator
+import com.heyzeusv.androidutilities.room.creators.RoomDataFileCreator
 
 class RoomProcessor(
     private val codeGenerator: CodeGenerator,
