+++ androidutilities-room/src/main/java/com/heyzeusv/androidutilities/room/creators/EntityFilesCreator.kt
@@ -1,12 +1,18 @@
-package com.heyzeusv.androidutilities.room
+package com.heyzeusv.androidutilities.room.creators
 
 import com.google.devtools.ksp.processing.CodeGenerator
 import com.google.devtools.ksp.processing.Dependencies
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSClassDeclaration
-import com.heyzeusv.androidutilities.room.csv.CsvData
-import com.heyzeusv.androidutilities.room.csv.CsvInfo
+import com.heyzeusv.androidutilities.room.CloseClass
+import com.heyzeusv.androidutilities.room.EmbeddedInfo
+import com.heyzeusv.androidutilities.room.EntityData
+import com.heyzeusv.androidutilities.room.FieldInfo
+import com.heyzeusv.androidutilities.room.PropertyInfo
+import com.heyzeusv.androidutilities.room.TypeConverterInfo
+import com.heyzeusv.androidutilities.room.util.CsvData
+import com.heyzeusv.androidutilities.room.util.CsvInfo
 import com.heyzeusv.androidutilities.room.util.addIndented
 import com.heyzeusv.androidutilities.room.util.asListTypeName
 import com.heyzeusv.androidutilities.room.util.containsNullableType
