--- androidutilities-room/src/main/java/com/heyzeusv/androidutilities/room/csv/CsvInfo.kt
@@ -1,10 +1,0 @@
-package com.heyzeusv.androidutilities.room.csv
-
-interface CsvInfo {
-    val csvFileName: String
-    val csvFieldToTypeMap: Map<String, String>
-}
-
-interface CsvData {
-    val csvRow: List<Any?>
-}
